fix(lap-564): resolve comments
