+++ src/components/templates/simulated-test-layout/header.tsx
@@ -1,9 +1,11 @@
 import { ChevronLeft } from "lucide-react";
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
+import { CallBackProps, STATUS } from "react-joyride";
 
 import ExitDialog from "@/components/organisms/simulated-test-dialog/exit-dialog";
 import StartDialog from "@/components/organisms/simulated-test-dialog/start-dialog";
+import SimulatedTestTour from "@/components/organisms/simulated-test-tour";
 import { Button } from "@/components/ui";
 import useCountdown from "@/hooks/use-countdown";
 
@@ -11,17 +13,39 @@ type HeaderProps = {
   currentPart: number;
 };
 export default function Header({ currentPart }: HeaderProps) {
+  const [run, setRun] = useState(false);
+  const [showStartDialog, setShowStartDialog] = useState(false);
   const { time, resume, isEnd } = useCountdown(60 * 40); // 40 minutes
   const { t } = useTranslation("simulatedTest");
+  useEffect(() => {
+    const isFirstTime = localStorage.getItem("simulatedTestFirstTime") !== "false";
+    if (isFirstTime) {
+      setRun(true);
+      localStorage.setItem("simulatedTestFirstTime", "false");
+    } else {
+      setShowStartDialog(true);
+    }
+  }, []);
   useEffect(() => {
     if (isEnd) {
       // TODO: submit the test
     }
   }, [isEnd]);
 
+  const handleTourCallback = (data: CallBackProps) => {
+    const { status } = data;
+    const finishedStatuses: string[] = [STATUS.FINISHED, STATUS.SKIPPED];
+
+    if (finishedStatuses.includes(status)) {
+      setRun(false);
+      resume();
+    }
+  };
+
   return (
     <>
-      <StartDialog onClose={() => resume()} />
+      {showStartDialog && <StartDialog onClose={() => resume()} />}
+      <SimulatedTestTour run={run} handleTourCallback={handleTourCallback} />
       <div className="grid w-full place-items-center border-b bg-white px-4 shadow-sm sm:h-20 sm:px-8">
         <div className="flex flex-col items-center">
           <h6 className="text-base font-bold uppercase sm:text-lg">
