+++ src/i18n/locales/en/simulated-test.json
@@ -59,7 +59,7 @@
             },
             "backAndNext": {
                 "title": "Back and Next Buttons",
-                "description": "Use the Back button to return to the previous question. Use the Next button to move to the next question."
+                "description": "Use the Back button to return to the previous part. Use the Next button to move to the next part."
             },
             "submit": {
                 "title": "Submit Button",
